+++ src/sim/sim_showrunner.py
@@ -1,6 +1,8 @@
 import pygame
 from src.controller import Controller
-from src.sim.dialog_scroll import DialogController
+from src.sim.dialog_controller import DialogController
+from src.sim.character import CharacterSprite
+from src.menu_element import MenuSprite
 
 
 class SimShowrunner(Controller):
@@ -17,7 +19,7 @@ class SimShowrunner(Controller):
         self.date_code = 0
         self.current_date_success = 0
         # sprite and menu box initialization
-        self.npc_house = CharacterSprite(f'../art/sim_sprites/{self.date_code}.png')
+        self.npc_house = CharacterSprite(self.date_code)
         # self.pc_house
         # self.player
         # always display static boxes
@@ -43,7 +45,7 @@ class SimShowrunner(Controller):
         self.sprites.add(self.boxes[self.STATICS])
         self.renderer = renderer
         self.renderer.set_background(self.date_code)
-        self.scroll = DialogController(f'../dialog/{self.date_code}.json')
+        self.scroll = DialogController(self.date_code)
         # self.running = True
 
     def begin(self):
@@ -121,48 +123,3 @@ class SimShowrunner(Controller):
                 r = [box.get_consequence() for box in boxes if box.get_rect().collidepoint(pygame.mouse.get_pos())]
                 return r if r != [] else None
         return None
-
-
-class MenuSprite(pygame.sprite.Sprite):
-    """
-    Any menu box item in the dating sim - dialog boxes, menu boxes, etc.
-    MenuBox instances may include a listener, or may be static elements on the screen.
-    """
-    def __init__(self, x, y, length, width, consequence=0, content=None, visible=True, image='default'):
-        super().__init__()
-        self.font = pygame.font.SysFont('Comic Sans MS', 30)
-        self.text_color = (0, 0, 0)
-        self.text_antialias = True
-        # self.image = pygame.image.load("path/to/element")
-        self.image = pygame.Surface((length, width))
-        self.image.fill("orange")
-        self.rect = self.image.get_rect()
-        self.rect.x = x
-        self.rect.y = y
-        self.content = self.font.render(content, self.text_antialias, self.text_color)
-        self.consequence = consequence
-
-    def get_rect(self):
-        return self.rect
-
-    def get_content(self):
-        return self.content
-
-    def get_content_pos(self):
-        return self.rect.x, self.rect.y
-
-    def get_consequence(self):
-        return self.consequence
-
-    def set_content(self, content):
-        self.content = self.font.render(content, False, (0, 0, 0))
-
-    def has_content(self):
-        return self.content is not None
-
-
-class CharacterSprite(pygame.sprite.Sprite):
-    def __init__(self, path):
-        super().__init__()
-        self.image = pygame.image.load(path)
-        self.rect = self.image.get_rect()
